backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
mysql-connector-python==8.4.0
werkzeug==3.0.3
